chef_converter: version constraint correctness
In older versions of the Chef client, we offered
the ability to specify multiple version
constraints for a cookbook dependency. We've since
removed that functionality and erchef and the new
depsolver do not support the array format returned
by the record. This patch coerces the data on the
way in to the database to prevent future errors.