Add fixup for old-style environments
At one early point, environments had a top-level "attributes" key. As
a first level fix, if such a key is present and its value is an empty
hash, then we simply delete it prior to validation and insert. Any
other value will fail validation and result in an error and failed
insert.