Add reindexing support for moser
This provides APIs for:
* selectively purging migrated data from the SOLR index, and
* inserting all index documents from a couch database file.

This includes some conversion code (derived from chef_object) to
normalized the stored Chef::* objects into the form expected by the
index.

This also includes a small refactor in moser_utils and moser_converter
to reduce repeated code when opening a couch database file and
processing it.