Coerce metadata.version to match version field in the cookbook_versions record.
We historically have not enforced any uniqueness the metadata.version field, or its correspondence with the version field of the record. The version field has been used for couchdb view computation and the like. 

This change sets the metadata.version field to the version field value, and uses that as value for the version column in sql.