Add support for ets tracking of transient migration state
Currently thi sis used to store a list of 'to be converted' users
in an ets table, and to allow the caller to retrieve users concurrently
with assurance that different caller processes will get different users.

Made this generic so that we can apply it for future data conversions in
which state is transient/we rely on the object data in the DB itself to track when
it has been converted in a persistent fashion.