Add lager support via script
In order to independently compile or package moser, lager must have the
parse_transform in erl_opts.  Because rebar does not support merging parent
erl_opts, we have an additional rebar.config.script that will search the
environment for a REBAR_PARENT.  If set, the script will merge the erl_opts of
the REBAR_PARENT config into the current config.  Precedence is given to the
current config.