Remove purge
The purge implementation requires implementing another callback.  However,
there is no delivery timeline for the purge work.  It should be relatively
trivial to restore the purge work and implement the callback once it is back
on the delivery timeline.

By deleting the purge implementation, we can remove a lot of the now redundant
apis in moser_state_tracker.